pyproject.toml
[tool.poetry]
name = "rag"
version = "0.1.0"
description = ""
authors = ["nevernever69 <nevernevernever69@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.2.11"
langchain-google-genai = "^1.0.8"
google-generativeai = "^0.7.2"
chromadb = "^0.5.5"
pypdf = "^4.3.1"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
langchain-community = "^0.2.10"
streamlit = "^1.37.0"
pdfplumber = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
